Add New activity Main and switch original Main Activity to GameActivity
