+++ app/src/main/res/layout/activity_game.xml
@@ -1,0 +1,79 @@
+<?xml version="1.0" encoding="utf-8"?>
+<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context=".MainActivity">
+
+    <ImageView
+        android:id="@+id/faceImage"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:src="@drawable/ic_cake"
+        android:layout_marginBottom="200dp"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintBottom_toBottomOf="parent"/>
+    <GridLayout
+        android:id="@+id/gridlayout"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginBottom="20dp"
+        android:rowCount="3"
+        android:columnCount="2"
+        android:layout_marginStart="30dp"
+        app:layout_constraintBottom_toTopOf="@id/cameraBtn"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toEndOf="parent">
+        <TextView
+            android:id="@+id/happiness"
+            android:layout_width="150dp"
+            android:layout_height="30dp"
+            android:text="@string/happiness" />
+        <TextView
+            android:id="@+id/sadness"
+            android:layout_width="150dp"
+            android:layout_height="30dp"
+            android:text="@string/sadness" />
+        <TextView
+            android:id="@+id/anger"
+            android:layout_width="150dp"
+            android:layout_height="30dp"
+            android:text="@string/anger"/>
+        <TextView
+            android:id="@+id/neutral"
+            android:layout_width="150dp"
+            android:layout_height="30dp"
+            android:text="@string/Neutral" />
+        <TextView
+            android:id="@+id/age"
+            android:layout_width="150dp"
+            android:layout_height="30dp"
+            android:text="@string/Age"/>
+        <TextView
+            android:id="@+id/gender"
+            android:layout_width="150dp"
+            android:layout_height="30dp"
+            android:text="@string/Gender"/>
+    </GridLayout>
+    <ProgressBar
+        android:id="@+id/loadingbar"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"/>
+    <Button
+        android:id="@+id/cameraBtn"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginBottom="30dp"
+        android:text="@string/Camera"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        />
+</android.support.constraint.ConstraintLayout>
