+++ app/src/main/res/layout/activity_main.xml
@@ -7,73 +7,33 @@
     tools:context=".MainActivity">
 
     <ImageView
-        android:id="@+id/faceImage"
+        android:id="@+id/titleImage"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:src="@drawable/ic_cake"
-        android:layout_marginBottom="200dp"
+        android:src="@drawable/lul"
+        android:layout_marginBottom="150dp"
+        android:scaleX="0.5"
+        android:scaleY="0.5"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintBottom_toBottomOf="parent"/>
-    <GridLayout
-        android:id="@+id/gridlayout"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginBottom="20dp"
-        android:rowCount="3"
-        android:columnCount="2"
-        android:layout_marginStart="30dp"
-        app:layout_constraintBottom_toTopOf="@id/cameraBtn"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent">
-        <TextView
-            android:id="@+id/happiness"
-            android:layout_width="150dp"
-            android:layout_height="30dp"
-            android:text="@string/happiness" />
-        <TextView
-            android:id="@+id/sadness"
-            android:layout_width="150dp"
-            android:layout_height="30dp"
-            android:text="@string/sadness" />
-        <TextView
-            android:id="@+id/anger"
-            android:layout_width="150dp"
-            android:layout_height="30dp"
-            android:text="@string/anger"/>
-        <TextView
-            android:id="@+id/neutral"
-            android:layout_width="150dp"
-            android:layout_height="30dp"
-            android:text="@string/Neutral" />
-        <TextView
-            android:id="@+id/age"
-            android:layout_width="150dp"
-            android:layout_height="30dp"
-            android:text="@string/Age"/>
-        <TextView
-            android:id="@+id/gender"
-            android:layout_width="150dp"
-            android:layout_height="30dp"
-            android:text="@string/Gender"/>
-    </GridLayout>
-    <ProgressBar
-        android:id="@+id/loadingbar"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"/>
     <Button
-        android:id="@+id/cameraBtn"
+        android:id="@+id/startBtn"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginBottom="30dp"
-        android:text="@string/Camera"
+        android:layout_marginBottom="80dp"
+        android:text="@string/Start"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         />
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="@string/GameTitle"
+        android:layout_marginBottom="40dp"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintBottom_toTopOf="@id/startBtn"/>
 </android.support.constraint.ConstraintLayout>
